FIX Errores Solucionados
