+++ client/src/components/Faq/scss/_component.scss
@@ -3,7 +3,7 @@
   background-repeat: no-repeat;
   box-shadow: 0px 50px 50px -20px rgba(53, 18, 122, 0.497159);
   border-radius: 23px;
-  width: 70%;
+  width: 60%;
   height: 80%;
   margin-bottom: 3%;
   display: flex;
