+++ client/src/components/Faq/scss/main.scss
@@ -1,9 +1,5 @@
 @import "config";
 
-h1 {
-  color: $page-color;
-}
-
 .container {
   width: 100%;
   height: 100%;
